% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_soils.R
\name{addApsoils}
\alias{addApsoils}
\title{Insert or Update soil into expDB}
\usage{
addApsoils(con, data, apsoil)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{A data frame includes all columns}

\item{apsoil}{Path to apsoil (Soils.soils)}
}
\description{
Insert or Update soil into expDB
}
