% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{averyLabel}
\alias{averyLabel}
\title{Generate labels for avery products}
\usage{
averyLabel(labels, file, product = "L7163")
}
\arguments{
\item{labels}{A character vector will be used to generate labels}

\item{file}{the file name to export}

\item{product}{The product ID of Avery}
}
\value{
No return values. Generated PDF file for all labels.
}
\description{
Generate labels for avery products
}
