% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkAlias}
\alias{checkAlias}
\title{Get index id by name.}
\usage{
checkAlias(name, alias = getAlias())
}
\arguments{
\item{name}{The old columns}

\item{alias}{The alias name for all columns}
}
\value{
The new name get rid of alias
}
\description{
Get index id by name.
}
