% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_trials.R
\name{dbAddDesigns}
\alias{dbAddDesigns}
\title{Add design for a trial}
\usage{
dbAddDesigns(con, data, extra_design = NULL)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{Trial design}

\item{extra_design}{The extra columns for design}
}
\value{
no return values
}
\description{
Add design for a trial
}
