% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_genotypes.R
\name{dbAddGenotype}
\alias{dbAddGenotype}
\title{Add genotypes into expDB}
\usage{
dbAddGenotype(con, genotypes)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{genotypes}{A string vector of genotypes}
}
\value{
No return values
}
\description{
Add genotypes into expDB
}
