% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_manager.R
\name{dbAddIrrigatons}
\alias{dbAddIrrigatons}
\title{Insert or Update irrigation into expDB}
\usage{
dbAddIrrigatons(con, data)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{A data frame includes all columns}
}
\value{
no return values
}
\description{
Insert or Update irrigation into expDB
}
