% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_log.R
\name{dbAddLog}
\alias{dbAddLog}
\title{Add log from expDB}
\usage{
dbAddLog(con, msg, date = format(Sys.time(), format = "\%Y-\%m-\%d"))
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{msg}{Add message into expdb}

\item{date}{Create time of message}
}
\value{
No return values
}
\description{
Add log from expDB
}
