% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_manager.R
\name{dbAddMethods}
\alias{dbAddMethods}
\title{Insert or Update methods into expDB}
\usage{
dbAddMethods(con, data)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{A data frame includes all columns}
}
\description{
Insert or Update methods into expDB
}
