% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_met.R
\name{dbAddMets}
\alias{dbAddMets}
\title{Insert and update met into expDB}
\usage{
dbAddMets(con, data)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{Met design}
}
\value{
no return values
}
\description{
Insert and update met into expDB
}
