% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_nodes.R
\name{dbAddNodes}
\alias{dbAddNodes}
\title{Add nodes into expDB}
\usage{
dbAddNodes(con, data)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{phenotype value}
}
\value{
no return values
}
\description{
Add nodes into expDB
}
