% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_phenotypes.R
\name{dbAddPhenotype}
\alias{dbAddPhenotype}
\title{Add design for a trial}
\usage{
dbAddPhenotype(con, data)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{phenotype value}
}
\value{
no return values
}
\description{
Add design for a trial
}
