% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_researcher.R
\name{dbAddResearcher}
\alias{dbAddResearcher}
\title{Insert and update researcher into expDB}
\usage{
dbAddResearcher(con, data)
}
\arguments{
\item{con}{A connection object as produced by dbConnect}

\item{data}{A data frame includes all columns}
}
\value{
no return values
}
\description{
Insert and update researcher into expDB
}
