% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_sites.R
\name{dbAddSites}
\alias{dbAddSites}
\title{Insert or update site into expDB}
\usage{
dbAddSites(con, data)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{A data frame includes all columns}
}
\value{
no return values
}
\description{
Insert or update site into expDB
}
