% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_source.R
\name{dbAddSource}
\alias{dbAddSource}
\title{Insert and update source into expDB}
\usage{
dbAddSource(con, data)
}
\arguments{
\item{con}{A connection object as produced by dbConnect}

\item{data}{A data frame includes all columns}
}
\value{
no return values
}
\description{
Insert and update source into expDB
}
