% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits.R
\name{dbAddTraits}
\alias{dbAddTraits}
\title{Insert or update trait into expDB}
\usage{
dbAddTraits(con, data)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{A data frame includes all columns}
}
\value{
no return values
}
\description{
Insert or update trait into expDB
}
