% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_trials.R
\name{dbAddTrials}
\alias{dbAddTrials}
\title{Insert or Update trial into expDB}
\usage{
dbAddTrials(con, data)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{A data frame includes all columns}
}
\value{
no return values
}
\description{
Insert or Update trial into expDB
}
