% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_trials.R
\name{dbAddTrialSoil}
\alias{dbAddTrialSoil}
\title{Add soil for a trial}
\usage{
dbAddTrialSoil(
  con,
  data,
  units = list(thickness = "cm", no3 = "kg/ha", nh4 = "kg/ha")
)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{Soil profiles for trials}

\item{units}{a list for the unit}
}
\value{
no return values
}
\description{
Add soil for a trial
}
