% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_met.R
\name{dbAddWeather}
\alias{dbAddWeather}
\title{Add weather records into expDB}
\usage{
dbAddWeather(con, data, name = NULL)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{data}{A string character for the path to met file,
a WeaAna object, or a data frame.}

\item{name}{The met name in the database 
if data is a data frame.}
}
\value{
no return values
}
\description{
Add weather records into expDB
}
