% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_db.R
\name{dbAppendTable}
\alias{dbAppendTable}
\title{Append a table into db and check the column name}
\usage{
dbAppendTable(con, table, data)
}
\arguments{
\item{con}{A connection object as produced by dbConnect}

\item{table}{The target table name}

\item{data}{A data frame to write into table}
}
\description{
Append a table into db and check the column name
}
