% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_export.R
\name{dbExportMet}
\alias{dbExportMet}
\title{Export trials weather records to met file}
\usage{
dbExportMet(con, output, na = NA, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{output}{The folder of output files}

\item{na}{The character for missing value with default NA}

\item{...}{All other arguments to define range of export trials.
All trials will be export if there are not arguments.
Supported arguments include trial (or trialcode)}
}
\value{
Write weather records into files. No return values.
}
\description{
Export trials weather records to met file
}
