% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_met.R
\name{dbGetDBName}
\alias{dbGetDBName}
\title{Get the database file name}
\usage{
dbGetDBName(con)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}
}
\description{
Get the database file name
}
