% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_manager.R
\name{dbGetFertilization}
\alias{dbGetFertilization}
\title{Get fertilization from database}
\usage{
dbGetFertilization(con, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{...}{Other arguments to specify meta data}
}
\value{
a data.frame for fertilization information
}
\description{
Get fertilization from database
}
