% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits_cal.R
\name{dbGetFieldPopulation}
\alias{dbGetFieldPopulation}
\title{Estimation of plant populations}
\usage{
dbGetFieldPopulation(con, trials = NULL, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{trials}{A data.frame to specify trials. If not NULL, other arguments
will be ignored.}

\item{...}{Arguments pass to dbGetTrials}
}
\value{
A data.frame for selected field population
}
\description{
Estimation of plant populations
}
