% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits_stem.R
\name{dbGetFieldTillerNumber}
\alias{dbGetFieldTillerNumber}
\title{Estimation of tiller number per unit area}
\usage{
dbGetFieldTillerNumber(con, trials = NULL, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{trials}{A data.frame to specify trials. If not NULL, other arguments
will be ignored.}

\item{...}{Arguments pass to dbGetTrials}
}
\value{
A data.frame for selected tiller number
}
\description{
Estimation of tiller number per unit area
}
