% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_genotypes.R
\name{dbGetGene}
\alias{dbGetGene}
\title{Get the gene information}
\usage{
dbGetGene(con)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}
}
\value{
a data.frame with all gene information
}
\description{
Get the gene information
}
