% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_genotypes.R
\name{dbGetGenotype}
\alias{dbGetGenotype}
\title{Get the genotype information}
\usage{
dbGetGenotype(con, name_only = FALSE)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{name_only}{Only return the name of genotypes}
}
\value{
data.frame with genotype information or a vector with genotype name 
if name_only = TRUE.
}
\description{
Get the genotype information
}
