% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_manager.R
\name{dbGetIrrigation}
\alias{dbGetIrrigation}
\title{Get irrigation from database}
\usage{
dbGetIrrigation(con, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{...}{Other arguments to specify meta data}
}
\value{
a data frame for irrigation information
}
\description{
Get irrigation from database
}
