% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_log.R
\name{dbGetLog}
\alias{dbGetLog}
\title{Get log from expDB}
\usage{
dbGetLog(con)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}
}
\value{
A data.frame with all logs
}
\description{
Get log from expDB
}
