% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_met.R
\name{dbGetMetInfo}
\alias{dbGetMetInfo}
\title{Get met information}
\usage{
dbGetMetInfo(con, name)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{name}{The met name}
}
\value{
a data.frame for met information
}
\description{
Get met information
}
