% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits_cal.R
\name{dbGetOrganFinalLeafNumber}
\alias{dbGetOrganFinalLeafNumber}
\title{Get the final leaf number}
\usage{
dbGetOrganFinalLeafNumber(con, trials = NULL, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{trials}{A data.frame to specify trials. If not NULL, other arguments}

\item{...}{Arguments to specific trials.}
}
\value{
A data.frame for selected final leaf number
}
\description{
The final leaf number is first retrived from trait
"O_FinalLeafNumber", then calculated from trait
"O_HaunIndex" if "O_FinalLeafNumber" is not observed.
Final leaf number equals the maximum value of "O_HaunIndex", which
should be an integer.
}
