% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits_cal.R
\name{dbGetOrganHaunIndex}
\alias{dbGetOrganHaunIndex}
\title{Get Haun Index}
\usage{
dbGetOrganHaunIndex(con, trials = NULL, avg = TRUE, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{trials}{A data.frame to specify trials. If not NULL, other arguments}

\item{avg}{Whether to calculate the average value}

\item{...}{Arguments to specific trials.}
}
\value{
A data.frame for selected Haun Index
}
\description{
The Haun Index is retrived from trait "O_HaunIndex", extending the final observationa
}
