% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_phenotypes.R
\name{dbGetPhenotype}
\alias{dbGetPhenotype}
\title{Get phenotype values through a group of conditions}
\usage{
dbGetPhenotype(
  con,
  traits = NULL,
  direction = "long",
  tt = FALSE,
  gene = FALSE,
  ...
)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{traits}{A list of traits. All traits will be returned if NULL}

\item{direction}{One of 'long' or 'wide' for reshape function}

\item{tt}{Whether to calculate thermal time}

\item{gene}{Whether to get gene information}

\item{...}{All other arguments to define range of export trials}
}
\value{
a data.frame for selected phenotypic values
}
\description{
Get phenotype values through a group of conditions
}
