% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_phenotypes.R
\name{dbGetPhenotype_}
\alias{dbGetPhenotype_}
\title{Get phenotype values through a group of conditions}
\usage{
dbGetPhenotype_(
  con,
  trials,
  traits = NULL,
  direction = "long",
  tt = FALSE,
  gene = FALSE
)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{trials}{Trials for filter}

\item{traits}{A list of traits. All traits will be returned if NULL}

\item{direction}{One of 'long' or 'wide' for reshape function}

\item{tt}{Whether to calculate thermal time}

\item{gene}{Whether to get gene information}
}
\description{
Get phenotype values through a group of conditions
}
