% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits_cal.R
\name{dbGetPlantStemElongation}
\alias{dbGetPlantStemElongation}
\title{Estimation of stem elongation}
\usage{
dbGetPlantStemElongation(con, trials = NULL, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{trials}{A data.frame to specify trials. If not NULL, other arguments
will be ignored.}

\item{...}{Arguments pass to dbGetTrials}
}
\value{
A data.frame for selected stem elongation stage
}
\description{
Estimation of stem elongation
}
