% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits_stem.R
\name{dbGetPlantTillerNumber}
\alias{dbGetPlantTillerNumber}
\title{Estimation of tiller number per plant}
\usage{
dbGetPlantTillerNumber(con, trials = NULL, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{trials}{A data.frame to specify trials. If not NULL, other arguments
will be ignored.}

\item{...}{Arguments pass to dbGetTrials}
}
\value{
A data.frame for selected tiller number for individual plant
}
\description{
Estimation of tiller number per plant
}
