% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_sites.R
\name{dbGetSites}
\alias{dbGetSites}
\title{Get site into expDB}
\usage{
dbGetSites(con)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}
}
\value{
a data.frame for all sites in the data base
}
\description{
Get site into expDB
}
