% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_source.R
\name{dbGetSource}
\alias{dbGetSource}
\title{Get source from expDB}
\usage{
dbGetSource(con)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}
}
\value{
A data.frame for all source in the data base
}
\description{
Get source from expDB
}
