% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits_stem.R
\name{dbGetStemTillerNumber}
\alias{dbGetStemTillerNumber}
\title{Estimation of stem/tiller number per unit area or plant}
\usage{
dbGetStemTillerNumber(con, trials, traits_order)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{trials}{A data.frame to specify trials}

\item{traits_order}{A list of traits for stem and tiller numbers.}
}
\description{
Estimation of stem/tiller number per unit area or plant
}
