% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits.R
\name{dbGetTraits}
\alias{dbGetTraits}
\title{Get trait list}
\usage{
dbGetTraits(con)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}
}
\value{
a data.frame for all traits in the data base
}
\description{
Get trait list
}
