% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits_cal.R
\name{dbGetTreatmentColumns}
\alias{dbGetTreatmentColumns}
\title{Get the unique columns for a treatment}
\usage{
dbGetTreatmentColumns(con, plot_id)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{plot_id}{id for plot}
}
\description{
Get the unique columns for a treatment
}
