% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_trials.R
\name{dbGetTrials}
\alias{dbGetTrials}
\title{Get trials by a groups of conditions.}
\usage{
dbGetTrials(con, design = TRUE, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{design}{Whether include design}

\item{...}{All other arguments to define range of export trials.
All trials will be export if there are not arguments.
Supported arguments include trial (or trialcode)}
}
\value{
A data.frame for selected trials
}
\description{
Get trials by a groups of conditions.
}
