% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_met.R
\name{dbGetWeather}
\alias{dbGetWeather}
\title{Get weather records from expDB}
\usage{
dbGetWeather(con, name, format = "data_frame", na = NA_character_, tz = "UTC")
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{name}{The met name}

\item{format}{The format of export dataset.}

\item{na}{The character for missing value with default NA}

\item{tz}{Time zone applied for hourly temperature}
}
\value{
a data.frame for all weather records
}
\description{
Get weather records from expDB
}
