% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_traits_cal.R
\name{dbGetZadoksStage}
\alias{dbGetZadoksStage}
\title{Obtain the key phenology stage}
\usage{
dbGetZadoksStage(con, trials, key_stage)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{trials}{A data.frame to specify trials. If not NULL, other arguments
will be ignored.}

\item{key_stage}{The key zadoks stage}
}
\value{
A data.frame for selected Zadoks stage
}
\description{
Obtain the key phenology stage
}
