% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_import.R
\name{dbImportXLSX}
\alias{dbImportXLSX}
\title{Import data from excel file}
\usage{
dbImportXLSX(con, xlsx, ignore_genotype = TRUE, ignore_trait = TRUE, ...)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{xlsx}{The path to excel file}

\item{ignore_genotype}{Ignore genotype tables when importing}

\item{ignore_trait}{Ignore trait table when importing}

\item{...}{Other arguments. Supported arguments include
\itemize{
   \item{extra_design: Extra columns in the experiment design.}
   \item{tz: The time zone for the hourly climates.}
}}
}
\value{
No return values
}
\description{
Import data from excel file
}
