% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_db.R
\name{dbInsertUpdateByRow}
\alias{dbInsertUpdateByRow}
\title{Insert new rows or update existing rows to a specific table according to a specific column (unique) by each row}
\usage{
dbInsertUpdateByRow(con, table, data, unique_col = "name")
}
\arguments{
\item{con}{A connection object as produced by dbConnect}

\item{table}{The target table name}

\item{data}{A data frame to write into table}

\item{unique_col}{A character vector to indentify each row in the table}
}
\description{
Insert new rows or update existing rows to a specific table according to a specific column (unique) by each row
}
