% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_trials.R
\name{dbListTrials}
\alias{dbListTrials}
\title{List all trials}
\usage{
dbListTrials(con)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}
}
\value{
A data.frame for all trials in the data base
}
\description{
List all trials
}
