% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{defaultMeasurement}
\alias{defaultMeasurement}
\title{Import a normal measurement}
\usage{
defaultMeasurement(con, phenotype, traits)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{phenotype}{Phenotype records}

\item{traits}{traits}
}
\description{
Import a normal measurement
}
