% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_db.R
\name{expdbConnect}
\alias{expdbConnect}
\title{Connect to expDB}
\usage{
expdbConnect(filename)
}
\arguments{
\item{filename}{The filename of SQLite}
}
\value{
a connection object as produced by dbConnect
}
\description{
Connect to expDB
}
\examples{
\dontrun{
con <- connect('filename')
}
}
