% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_db.R
\name{expdbCreateDB}
\alias{expdbCreateDB}
\title{create to expDB}
\usage{
expdbCreateDB(filename, system_traits = TRUE)
}
\arguments{
\item{filename}{The filename of new expDB}

\item{system_traits}{Whether to import system traits}
}
\value{
a connection object as produced by dbConnect
}
\description{
create to expDB
}
