% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_db.R
\name{expdbDisconnect}
\alias{expdbDisconnect}
\title{Didconnect to expDB}
\usage{
expdbDisconnect(con)
}
\arguments{
\item{con}{a connection object as produced by dbConnect.}
}
\value{
no return values
}
\description{
Didconnect to expDB
}
\examples{
\dontrun{
con <- connect('filename')
disconnect(con)
}
}
