% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generateName}
\alias{generateName}
\title{Generate unique name with digest method}
\usage{
generateName(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Generate unique name with digest method
}
