% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_db.R
\name{getIdByUniqueIndex}
\alias{getIdByUniqueIndex}
\title{Get index id by unique_columns.}
\usage{
getIdByUniqueIndex(
  con,
  table,
  data,
  unique_col = "name",
  data_col = unique_col,
  ignore_case = FALSE
)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{table}{the table name}

\item{data}{A data frame to write into table}

\item{unique_col}{A character vector to identify each row in the table}

\item{data_col}{The column names in the data.frame}

\item{ignore_case}{Whether ignore_case}
}
\value{
id of unique_col
}
\description{
Get index id by unique_columns.
}
