% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvestQuadratDetail}
\alias{harvestQuadratDetail}
\title{Process quadrat (detail) harvestQuadratDetail}
\usage{
harvestQuadratDetail(con, records)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{records}{Phenotype records}
}
\value{
no return values
}
\description{
Process quadrat (detail) harvestQuadratDetail
}
