% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{measurementLeafSize}
\alias{measurementLeafSize}
\title{Import the measurement of leaf size}
\usage{
measurementLeafSize(con, phenotype)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{phenotype}{Phenotype records}
}
\description{
Import the measurement of leaf size
}
