% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_soils.R
\name{parseApsoil}
\alias{parseApsoil}
\title{parse APSIM soil file}
\usage{
parseApsoil(data, apsoil)
}
\arguments{
\item{data}{the data object}

\item{apsoil}{file path to apsoil}
}
\description{
parse APSIM soil file
}
