% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{yearDay2Date}
\alias{yearDay2Date}
\title{Convert year and day of year to date}
\usage{
yearDay2Date(day, year)
}
\arguments{
\item{day}{day of year}

\item{year}{year}
}
\description{
Convert year and day of year to date
}
