% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_genotypes.R
\name{dbAddGene}
\alias{dbAddGene}
\title{Add gene information into database}
\usage{
dbAddGene(con, genes)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{genes}{A data.frame of genes}
}
\value{
No return values
}
\description{
Add gene information into database
}
