% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdb_genotypes.R
\name{dbGenotypeCheckName}
\alias{dbGenotypeCheckName}
\title{Check genotype names}
\usage{
dbGenotypeCheckName(con, genotype)
}
\arguments{
\item{con}{a connection object as produced by dbConnect}

\item{genotype}{The genotype name will be checked}
}
\value{
A vector with check genotype names
}
\description{
Check genotype names
}
